tests/wpt/web-platform-tests/html/canvas/offscreen/drawing-images-to-the-canvas/2d.drawImage.svg.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.drawImage.svg</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.drawImage.svg</h1>
<p class="desc">drawImage() of an SVG image</p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var promise = new Promise(function(resolve, reject) {
    var xhr = new XMLHttpRequest();
    xhr.open("GET", '/images/green.svg');
    xhr.responseType = 'blob';
    xhr.send();
    xhr.onload = function() {
      resolve(xhr.response);
    };
  });
  promise.then(function(response) {
    createImageBitmap(response).then(bitmap => {
      ctx.drawImage(bitmap, 0, 0);
      _assertPixelApprox(canvas, 50,25, 0,255,0,255, 2);
    });
  });

}, "drawImage() of an SVG image");
</script>
